refactor(SF-119): split create_app — admin endpoints + bootstrap helpers
core/app.py was a 406-line factory doing five jobs: config resolution,
registry init, lifespan, request middleware, eight admin endpoints,
plus three large schema-walker helpers.

Split it:

- core/admin_router.py (new, 249 lines): register_admin_routes(app)
  attaches /health, /system/processes, /plugins, /plugins/{name}/...,
  /config/validate. Also owns the schema helpers (_inject_tag_enums,
  _collapse_nullable, _inject_examples, _get_process_names) since
  only the admin router consumes them.
- app.py: now 147 lines (-64%). create_app is ~50 lines orchestrating
  three private helpers:
  - _resolve_config: arg → pending → env var → sf.json
  - _build_lifespan: app.startup/shutdown + subprocess ready event
  - _activate_plugins: discover, activate configured set, cleanup
    stale side effects of inactive ones

Behavior unchanged. 706 plugin unit tests pass.

https://app.asana.com/1/1185126988600652/project/1213628819033917/task/SF-119
